requirements.txt
pandas
pymongo
Flask-SQLAlchemy
textblob
finnhub_python